removed etag check for posts (not in 1.0.2)-removed tests that tested it

+++ lrs/managers/AgentProfileManager.py
@@ -14,41 +14,68 @@ class AgentProfileManager():
     def __init__(self, agent):
     	self.Agent = agent
 
-    @transaction.commit_on_success        
-    def post_profile(self, request_dict):
-        post_profile = request_dict['profile']
-        profile_id = request_dict['params']['profileId']
-
-        p, created = AgentProfile.objects.get_or_create(profileId=profile_id,agent=self.Agent)
+    def save_non_json_profile(self, p, profile, request_dict):
+        p.content_type = request_dict['headers']['CONTENT_TYPE']
+        p.etag = etag.create_tag(profile.read())
         
-        if created:
-            p.json_profile = post_profile
-            p.content_type = request_dict['headers']['CONTENT_TYPE']
-            p.etag = etag.create_tag(post_profile)
-
-            if 'headers' in request_dict and ('updated' in request_dict['headers'] and request_dict['headers']['updated']):
-                p.updated = request_dict['headers']['updated']
-            else:
-                p.updated = datetime.datetime.utcnow().replace(tzinfo=utc)
+        if 'updated' in request_dict['headers'] and request_dict['headers']['updated']:
+            p.updated = request_dict['headers']['updated']
         else:
-            etag.check_preconditions(request_dict,p, required=True)
-            orig_prof = json.loads(p.json_profile)
-            post_profile = json.loads(post_profile)
-            if not isinstance(post_profile, dict):
-                raise ParamError("The document was not able to be parsed into a JSON object.")
-            else:
-                merged = json.dumps(dict(orig_prof.items() + post_profile.items()))
-            p.json_profile = merged
-            p.etag = etag.create_tag(merged)
             p.updated = datetime.datetime.utcnow().replace(tzinfo=utc)
+        
+        # Go to beginning of file
+        profile.seek(0)
+        fn = "%s_%s" % (p.agent_id, request_dict.get('filename', p.id))
+        p.profile.save(fn, profile)
 
         p.save()
 
+    @transaction.commit_on_success        
+    def post_profile(self, request_dict):
+        # get/create profile
+        p, created = AgentProfile.objects.get_or_create(profileId=request_dict['params']['profileId'],agent=self.Agent)
+
+        if "application/json" not in request_dict['headers']['CONTENT_TYPE']:
+            try:
+                post_profile = ContentFile(request_dict['profile'].read())
+            except:
+                try:
+                    post_profile = ContentFile(request_dict['profile'])
+                except:
+                    post_profile = ContentFile(str(request_dict['profile']))            
+            self.save_non_json_profile(p, post_profile, request_dict)
+        else:
+            post_profile = request_dict['profile']
+            # If incoming profile is application/json and if a profile didn't already exist with the same agent and profileId
+            if created:
+                p.json_profile = post_profile
+                p.content_type = request_dict['headers']['CONTENT_TYPE']
+                p.etag = etag.create_tag(post_profile)
+            # If incoming profile is application/json and if a profile already existed with the same agent and profileId 
+            else:
+                orig_prof = json.loads(p.json_profile)
+                post_profile = json.loads(post_profile)
+                if not isinstance(post_profile, dict):
+                    raise ParamError("The document was not able to be parsed into a JSON object.")
+                else:
+                    # json.dumps changes the format of the string rep of the dict
+                    merged = json.dumps(dict(orig_prof.items() + post_profile.items()))
+                p.json_profile = merged
+                p.etag = etag.create_tag(merged)
+            
+            #Set updated
+            if 'updated' in request_dict['headers'] and request_dict['headers']['updated']:
+                p.updated = request_dict['headers']['updated']
+            else:
+                p.updated = datetime.datetime.utcnow().replace(tzinfo=utc)
+            p.save()
+
     @transaction.commit_on_success
     def put_profile(self, request_dict):
-        profile_id = request_dict['params']['profileId']
-        p,created = AgentProfile.objects.get_or_create(profileId=profile_id,agent=self.Agent)
+        # get/create profile
+        p, created = AgentProfile.objects.get_or_create(profileId=request_dict['params']['profileId'],agent=self.Agent)
 
+        # Profile being PUT is not json
         if "application/json" not in request_dict['headers']['CONTENT_TYPE']:
             try:
                 profile = ContentFile(request_dict['profile'].read())
@@ -57,43 +84,34 @@ class AgentProfileManager():
                     profile = ContentFile(request_dict['profile'])
                 except:
                     profile = ContentFile(str(request_dict['profile']))
-        
 
+            # If a profile already existed with the profileId and activityId
             if not created:
-                etag.check_preconditions(request_dict,p, required=True)
-                try:
-                    p.profile.delete()
-                except OSError:
-                    # p was probably json before.. gotta clear that field
-                    p.json_profile = {}
-            self.save_profile(p, created, profile, request_dict)
+                #If it already exists delete it
+                etag.check_preconditions(request_dict, p, required=True)
+                if p.profile:
+                    try:
+                        p.profile.delete()
+                    except OSError:
+                        # probably was json before
+                        p.json_profile = {}
+            self.save_non_json_profile(p, profile, request_dict)
+        # Profile being PUT is json
         else:
+            # If a profile already existed with the profileId and activityId (overwrite existing profile data)
             if not created:
                 etag.check_preconditions(request_dict, p, required=True)
             the_profile = request_dict['profile']
             p.json_profile = the_profile
             p.content_type = request_dict['headers']['CONTENT_TYPE']
             p.etag = etag.create_tag(the_profile)
-
-            if 'headers' in request_dict and ('updated' in request_dict['headers'] and request_dict['headers']['updated']):
+            
+            #Set updated
+            if 'updated' in request_dict['headers'] and request_dict['headers']['updated']:
                 p.updated = request_dict['headers']['updated']
             else:
                 p.updated = datetime.datetime.utcnow().replace(tzinfo=utc)
             p.save()
-
-    def save_profile(self, p, created, profile, request_dict):
-        p.content_type = request_dict['headers']['CONTENT_TYPE']
-        p.etag = etag.create_tag(profile.read())
-        if 'headers' in request_dict and ('updated' in request_dict['headers'] and request_dict['headers']['updated']):
-            p.updated = request_dict['headers']['updated']
-        else:
-            p.updated = datetime.datetime.utcnow().replace(tzinfo=utc)
-        profile.seek(0)
-        if created:
-            p.save()
-
-        fn = "%s_%s" % (p.agent_id,request_dict.get('filename', p.id))
-        p.profile.save(fn, profile)
     
     def get_profile(self, profile_id):
         try:
@@ -107,14 +125,10 @@ class AgentProfileManager():
         if since:
             try:
                 # this expects iso6801 date/time format "2013-02-15T12:00:00+00:00"
-                profs = self.Agent.agentprofile_set.filter(updated__gte=since)
+                profs = self.Agent.agentprofile_set.filter(updated__gt=since)
             except ValidationError:
                 err_msg = 'Since field is not in correct format for retrieval of agent profiles'
-                raise ParamError(err_msg) 
-            except:
-                err_msg = 'There are no profiles associated with the id'
-                raise IDNotFoundError(err_msg) 
-
+                raise ParamError(err_msg)  
             ids = [p.profileId for p in profs]
         else:
             ids = self.Agent.agentprofile_set.values_list('profileId', flat=True)
@@ -122,11 +136,9 @@ class AgentProfileManager():
 
     def delete_profile(self, profileId):
         try:
-            prof = self.get_profile(profileId)
-            prof.delete()
+            self.get_profile(profileId).delete()
+        # we don't want it anyway
         except AgentProfile.DoesNotExist:
-            pass #we don't want it anyway
+            pass 
         except IDNotFoundError:
             pass
-        except OSError:
-            pass # this is ok,too
